requirements.txt
# Minimal public dependencies
pandas
openpyxl
pyyaml
tqdm
numpy
torch
transformers
